refactor: split webserver routes into separate modules
